structured_outputs/requirements.txt
azure.identity
openai
rick
pydantic